pyproject.toml
[project]
name = "ldm"
version = "0.1.0"
requires-python = ">=3.8,<3.9"
dependencies = [
    "numpy==1.23.1",
    "torch==2.0.1",
    "torchvision==0.15.2",
    "albumentations==1.3.0",
    "opencv-python==4.6.0.66",
    "imageio==2.9.0",
    "imageio-ffmpeg==0.4.2",
    "pytorch-lightning==1.4.2",
    "omegaconf==2.1.1",
    "test-tube>=0.7.5",
    "streamlit==1.12.1",
    "einops==0.3.0",
    "transformers==4.19.2",
    "webdataset==0.2.5",
    "kornia==0.6",
    "open-clip-torch==2.0.2",
    "invisible-watermark>=0.1.5",
    "streamlit-drawable-canvas==0.8.0",
    "torchmetrics==0.6.0",
    "python-dotenv",
    "wandb",
    "datasets>=3.1.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
